Completed The signup page and improved the login page functionality
