+++ src/style.css
@@ -16,5 +16,6 @@
 }
 .svg{
     position: absolute;
+    margin-top: 6rem;
 }
